requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
gunicorn>=21.2
dash==2.17.*
plotly>=5.22
pandas>=2.2
pyarrow>=16.1
deltalake>=0.19
fsspec>=2024.6.0
adlfs>=2024.4.1
azure-identity>=1.16.0
azure-storage-file-datalake>=12.15.0
azure-storage-blob
httpx>=0.27
pydantic>=2.8
python-dotenv>=1.0
pydantic
pydantic-settings
